Use correct _STAT_VER
